// Portfolio/src/app/interfaces/iprofile.ts
import { ILocation } from "./ilocation";

export interface IProfile {
  profileId?: any;
  firstName?: string;
  secondName?: string;
  lastName?: string;
  locationId?: any;
  location?: any;
  statusId?: any;
  status?: any;
  descripTion?: string;
  contactId?: any;
  contact?: any;
  birthDay?: any;
}